Support custom square backgrounds on board rendering
Eyes have to ignore changes in chessboard background if the position isn't changed. To achieve this goal, first of all BoardRenderer has to be able to change squares' background with arbitrary colours.

A fluent interface has been added to BoardRendered which allows clients to specify the background for a specific square. Some tests were defined to assure the correct implementation.

Completes: #24
See also: #23