Implement Algebraic Notation
To describe chess moves and let Chesslave's mouth announce them, the most widely used chess notation was added to the model to represent a move. Standard version of the Algebraic Notation was implemented, but the design can be easily extended to add other Algebraic notations.

Warning: ambiguous moves are currently not recognised. Rules have to be fixed.

See: #7